Create a named list of connection arguments to translate to bcp and sqlcmd options

Create a named list of connection arguments to translate to bcp and sqlcmd options

makeConnectArgs(
  server,
  database,
  username,
  password,
  trustedconnection = TRUE,
  trustservercert = FALSE,
  azure = FALSE
)

Arguments

server

the instance of SQL Server to which to connect

database

specifies the database to connect to

username

login ID

password

password for login ID

trustedconnection

use integrated security, username and password are not required

trustservercert

trust the server certificate

azure

use Azure Active Directory authentication, does not work with integrated authentication.

Value

a list with connection arguments